+++ inst/tailwind/tailwind.config.js
@@ -1,0 +1,11 @@
+module.exports = {
+  purge: ['./srcjs/*.js'],
+  darkMode: false,
+  theme: {
+    extend: {},
+  },
+  variants: {
+  extend: {},
+  },
+  plugins: [],
+};